requirements.txt
cx_Oracle==8.3.0
google-cloud-bigquery==2.28.1
google-auth

#pip install -r requirements.txt
